backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
faker==22.0.0